feat : 공통 기능 분리 - closes #22
pagination
취소 버튼 기능
