+++ src/main/resources/templates/booking/history.html
@@ -5,8 +5,9 @@
     <meta charset="UTF-8">
     <link href="/css/booking/history.css" rel="stylesheet">
     <script src="/js/booking/api.js"></script>
-    <script src="/js/booking/format.js"></script>
-    <script src="/js/booking/history.js" defer></script>
+    <script src="/js/booking/utils/format.js"></script>
+    <script src="/js/booking/utils/pagination.js"></script>
+    <script type="module" src="/js/booking/history.js" defer></script>
     <title>회의실 예약 내역</title>
 </head>
 <body>
