+++ src/main/resources/templates/booking/statistics.html
@@ -4,7 +4,7 @@
 <head>
   <meta charset="UTF-8">
   <script src="/js/booking/api.js"></script>
-  <script src="/js/booking/format.js"></script>
+  <script src="/js/booking/utils/format.js"></script>
   <script src="/js/booking/statistics.js" defer></script>
   <title>Title</title>
 </head>
