+++ src/main/resources/templates/booking/success.html
@@ -5,7 +5,7 @@
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css">
     <link href="/css/booking/success.css" rel="stylesheet"/>
     <script src="/js/booking/api.js"></script>
-    <script src="/js/booking/format.js"></script>
+    <script src="/js/booking/utils/format.js"></script>
     <script src="/js/booking/success.js"></script>
     <title>예약 성공</title>
 </head>
